+++ ExpandWorldPrefabs/InfoManager.cs
@@ -1,5 +1,9 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using ExpandWorldData;
+using HarmonyLib;
+using UnityEngine;
 
 namespace ExpandWorld.Prefab;
 
@@ -19,7 +23,6 @@ public class InfoManager
   public static readonly PrefabInfo RemoveDatas = new();
   public static readonly PrefabInfo RepairDatas = new();
   public static readonly PrefabInfo DamageDatas = new();
-  public static readonly PrefabInfo TargetDatas = new();
   public static readonly PrefabInfo StateDatas = new();
   public static readonly PrefabInfo CommandDatas = new();
   public static readonly PrefabInfo SayDatas = new();
@@ -30,7 +33,6 @@ public class InfoManager
     RemoveDatas.Clear();
     RepairDatas.Clear();
     DamageDatas.Clear();
-    TargetDatas.Clear();
     StateDatas.Clear();
     CommandDatas.Clear();
     SayDatas.Clear();
@@ -40,6 +42,17 @@ public class InfoManager
     Select(info.Type).Add(info);
     if (info.Type == ActionType.Say)
       Select(ActionType.Command).Add(info);
+  }
+  public static void Patch()
+  {
+    EWP.Harmony.UnpatchSelf();
+    EWP.Harmony.PatchAll();
+    if (CreateDatas.Exists)
+      HandleCreated.Patch(EWP.Harmony);
+    if (RemoveDatas.Exists)
+      HandleDestroyed.Patch(EWP.Harmony);
+    if (RepairDatas.Exists || DamageDatas.Exists || StateDatas.Exists || CommandDatas.Exists || SayDatas.Exists)
+      HandleRPC.Patch(EWP.Harmony);
   }
 
 
@@ -60,7 +73,32 @@ public class InfoManager
     if (IsCreatureCache.TryGetValue(prefab, out var result))
       return result;
     return IsCreatureCache[prefab] = ZNetScene.instance.GetPrefab(prefab)?.GetComponent<BaseAI>() != null;
+  }
 
+  private static Dictionary<string, int> PrefabCache = [];
+  public static IEnumerable<int> GetPrefabs(string prefab)
+  {
+    var p = prefab.ToLowerInvariant();
+    if (PrefabCache.Count == 0)
+      PrefabCache = ZNetScene.instance.m_namedPrefabs.ToDictionary(pair => pair.Value.name, pair => pair.Key);
+    if (p == "*")
+      return PrefabCache.Values;
+    if (p[0] == '*' && p[p.Length - 1] == '*')
+    {
+      p = p.Substring(1, p.Length - 2);
+      return PrefabCache.Where(pair => pair.Key.ToLowerInvariant().Contains(p)).Select(pair => pair.Value);
+    }
+    if (p[0] == '*')
+    {
+      p = p.Substring(1);
+      return PrefabCache.Where(pair => pair.Key.EndsWith(p, StringComparison.OrdinalIgnoreCase)).Select(pair => pair.Value);
+    }
+    if (p[p.Length - 1] == '*')
+    {
+      p = p.Substring(0, p.Length - 1);
+      return PrefabCache.Where(pair => pair.Key.StartsWith(p, StringComparison.OrdinalIgnoreCase)).Select(pair => pair.Value);
+    }
+    return [];
   }
 }
 
@@ -68,57 +106,52 @@ public class PrefabInfo
 {
   private static readonly int CreatureHash = "creature".GetStableHashCode();
   public readonly Dictionary<int, List<Info>> Prefabs = [];
-  public readonly List<Info> WildPrefabs = [];
   public readonly List<Info> Creatures = [];
+  public bool Exists => Prefabs.Count > 0 || Creatures.Count > 0;
 
 
   public void Clear()
   {
     Prefabs.Clear();
     Creatures.Clear();
-    WildPrefabs.Clear();
   }
   public void Add(Info info)
   {
-    if (info.Prefab.Contains("*"))
+    var prefabs = DataManager.ToList(info.Prefabs);
+    HashSet<int> hashes = [];
+    foreach (var prefab in prefabs)
     {
-      WildPrefabs.Add(info);
-      return;
+      if (prefab.Contains("*"))
+        hashes.UnionWith(InfoManager.GetPrefabs(prefab));
+      else
+        hashes.Add(prefab.GetStableHashCode());
     }
-    var hash = info.Prefab.GetStableHashCode();
-    if (hash == CreatureHash)
-      Creatures.Add(info);
-    else
+    foreach (var hash in hashes)
     {
-      if (!Prefabs.TryGetValue(hash, out var list))
-        Prefabs[hash] = list = [];
-      list.Add(info);
+      if (hash == CreatureHash)
+        Creatures.Add(info);
+      else
+      {
+        if (!Prefabs.TryGetValue(hash, out var list))
+          Prefabs[hash] = list = [];
+        list.Add(info);
+      }
     }
   }
 
   public bool TryGetValue(int prefab, out List<Info> list)
   {
     var ret = Prefabs.TryGetValue(prefab, out list);
-    var wild = GetWildPrefabs(prefab);
     var creatures = GetCreaturePrefabs(prefab);
-    if (!ret && wild == null && creatures == null)
+    if (!ret && creatures == null)
       return false;
     if (ret)
     {
-      if (wild != null && creatures != null)
-        list = [.. list, .. wild, .. creatures];
-      else if (wild != null)
-        list = [.. list, .. wild];
-      else if (creatures != null)
+      if (creatures != null)
         list = [.. list, .. creatures];
-      return true;
     }
-    if (creatures != null && wild != null)
-      list = [.. wild, .. creatures];
     else if (creatures != null)
       list = creatures;
-    else if (wild != null)
-      list = wild;
     return true;
   }
   private List<Info>? GetCreaturePrefabs(int prefab)
@@ -130,14 +163,4 @@ public class PrefabInfo
       return null;
     return Creatures;
   }
-  private List<Info>? GetWildPrefabs(int prefab)
-  {
-    if (WildPrefabs.Count == 0) return null;
-    var obj = ZNetScene.instance.GetPrefab(prefab);
-    if (!obj) return null;
-    var name = obj.name;
-    var ret = WildPrefabs.Where(info => Helper2.CheckWild(info.Prefab, name)).ToList();
-    return ret.Count == 0 ? null : ret;
-  }
-
 }